READ.me
The dataset is all about the amount of people that committed suicide in some particular country. In the notebook,three models are used, they are linear regression model which performed not the expectation , decision tree which preformed really well but the random forest model did really than the two former ones with almost 99% accuracy. THe EDA and data cleaning and processing is done also in the notebook